pyproject.toml
[project]
name = "vector_search"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = ["numpy"]

[project.optional-dependencies]
hw1 = [
    "scikit-learn",
    "matplotlib",
    "umap-learn"
]

hw2 = [
  "annoy",
  "faiss-cpu",
  "hnswlib",
  "numpy",
  "orjson"
]

hw3 = [
  "qdrant-client[grpc]",
  "orjson"
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"